fix(widget): queue follow-ups until the active reply settles
