+++ .changeset/widget-follow-up-order.md
@@ -1,0 +1,8 @@
+---
+'@opencx/widget-core': patch
+'@opencx/widget-react-headless': patch
+'@opencx/widget-react': patch
+'@opencx/widget': patch
+---
+
+Keep follow-up messages queued until the current reply finishes. Prevent the previous reply from briefly appearing twice when the next reply starts.
